Compute not-cached intervals on file open
This is probably a huge memory saver, and pushes the computation
complexity in the open() part, not the close() part.

Instead of storing the whole info vector from mincore, we store it
temporarily, compute which parts of the file are not cached, and then
free the vector again, retaining only a singly linked list of byte
ranges we’ll want to fadvise() on when closing the file.

For example, consider an MP3 file which I played at 0:00, skipped to
10:00, then skipped to 20:00. It has three small cached regions (plus
the last page is cached), hence we store three “uncached intervals”.
On close(), these intervals will be marked “don’t need”. Additionally,
if the file size changes, everything newly added will be marked “don’t
need”, as happens with the new file (fd = 4).

    $ ./nocache -D /dev/stderr cp ~/musik/fernanda.mp3 /tmp/x
    [nocache] DEBUG: Executing: cp /home/feh/musik/fernanda.mp3 /tmp/x
    [nocache] DEBUG: open(pathname=/home/feh/musik/fernanda.mp3, flags=0x0, mode=00) = 3
    [nocache] DEBUG: fd_get_pageinfo(fd=3): st.st_size=80695431, nr_pages=19702
    [nocache] DEBUG: store_pageinfo(fd=3): pages in cache: 183/19702 (0.9%)  [filesize=78804.1K, pagesize=4K]
    [nocache] DEBUG: open(pathname=/tmp/x, flags=0x201, mode=00) = 4
    [nocache] DEBUG: fd_get_pageinfo(fd=4): st.st_size=0, nr_pages=0
    [nocache] DEBUG: store_pageinfo(fd=4): pages in cache: 0/0 (0.0%)  [filesize=0.0K, pagesize=4K]
    [nocache] DEBUG: fadv_dontneed(fd=4, from=0, len=0 [till new end, file has grown])
    [nocache] DEBUG: close(4)
    [nocache] DEBUG: fadv_dontneed(fd=3, from=245760, len=14266368)
    [nocache] DEBUG: fadv_dontneed(fd=3, from=14761984, len=14258176)
    [nocache] DEBUG: fadv_dontneed(fd=3, from=29270016, len=51425280)
    [nocache] DEBUG: close(3)

It might be beneficial to “chunk” the mincore() call, since one call
needs filesize/PAGESIZE bytes of memory, so we can fit into 16MB of heap
memory only the map for a file of size 16*1024^2*4096/1024^3 = 64 GB.
On terabyte-sized file, we would allocate a tremendous amount of memory,
and by repeatedly calling mincore() on 64GB-chunks of the file and
reusing the allocated heap space, we could probably improve performance
and waste less space. This is for a future commit.