Use two-tier hierarchy of locks for fast yet safe access to fd info array
The previous approach relied on a global lock, but acquired it as
shortly as possible to avoid contention. Since there was no easy way to
transfer lock ownership between function calls the lock had to be
released before being acquired again by some of the called functions.
There was at least one race condition where this was not done in a safe
manner, and making it safe would have required holding the lock a *very*
long time, likely leading to a performance regression.

The new approach is a bit more complex because it uses locks per fd
(i.e. much finer granularity so we don’t hold a global lock while doing
file system I/O) but does so in a safe manner. The new approach is
described in an inline code comment, reproduced here:

    /* Info about a file descriptor 'fd' is stored in fds[fd]. Before
     * accessing an element, callers MUST both check fds_lock != NULL
     * and then acquire fds_lock[fd] while holding the fds_iter_lock.
     * While any mutex in fds_lock is held, fds_lock will not be freed
     * and set to NULL. */

This should fix GitHub issue #30.

Hindsight is 20/20: Storing the file descriptor n in fds[m] where
usually n != m is not the most ideal thing to do. So I drive-by fixed
that: Now info about fd n is always stored in fds[n].