Fix off by one issue with NOCACHE_MAX_FDS
Since NOCACHE_MAX_FDS should be an artificial cap for RLIMIT_NOFILE, it
should follow the same convention for its value - it should be one
greater than the maximum file descriptor that would be handled by
nocache. However, it currently is 2 greater than that file descriptor.

This commit resolves that. Also:
- tweaked mafd.t so it will correctly check for this problem
- touched the README file with small correction